paralelize page data fetching
