+++ web/src/app/[slug]/@issues/loading.tsx
@@ -1,0 +1,9 @@
+import Spinner from "@/components/spinner"
+
+export default function Loading() {
+  return(
+    <div className="flex flex-col items-center justify-center min-h-screen">
+      <Spinner />
+    </div>
+  )
+}
