+++ web/src/app/[slug]/@issues/types.ts
@@ -1,0 +1,5 @@
+export interface IssuesProps {
+  params: {
+    slug: string
+  }
+}
