+++ web/src/app/[slug]/@product/types.ts
@@ -1,0 +1,5 @@
+export interface ProductProps {
+  params: {
+    slug: string
+  }
+}
